+++ src/corral/backend/base.py
@@ -1,15 +1,15 @@
 import time
 from abc import ABC, abstractmethod
-from collections.abc import Callable
 from copy import deepcopy
 from dataclasses import dataclass, field
 from datetime import datetime, timezone
-from enum import Enum, StrEnum
+from enum import StrEnum
 from pathlib import Path
 from typing import Any
 
 from loguru import logger
-from pydantic import BaseModel
+
+from corral.backend.tool import Tool, ToolCallStatus
 
 
 class Role(StrEnum):
@@ -18,30 +18,6 @@ class Role(StrEnum):
     AGENT = "agent"
     ENVIRONMENT = "environment"
     TOOL = "tool"
-
-
-class ToolCallStatus(Enum):
-    SUCCESS = "success"
-    INVALID_TOOL = "invalid_tool"
-    INVALID_ARGS = "invalid_args"
-    EXECUTION_ERROR = "execution_error"
-
-
-class ToolRequest(BaseModel):
-    tool_name: str
-    arguments: dict[str, Any]
-
-
-@dataclass
-class ToolArgument:
-    """Specification for a tool argument"""
-
-    name: str
-    type: str
-    description: str
-    required: bool = True
-    default: Any = None
-    choices: list[Any] | None = None  # from transformers
 
 
 @dataclass
@@ -67,7 +43,7 @@ class LLMMessage:
 @dataclass
 class TaskState:
     task_id: str
-    task_prompt: str
+    task_prompt: str | list[dict]
     trial_id: str = "0"
     messages: list[LLMMessage] = field(default_factory=list)
     tool_calls: list[ToolCall] = field(default_factory=list)
@@ -101,94 +77,6 @@ class TaskState:
         if self.end_time and self.start_time:
             return (self.end_time - self.start_time).total_seconds()
         return None
-
-
-class Tool:
-    """Base class for tools
-    Inherit from this class to create new tools.
-    Should have an execute method that performs the tool's functionality.
-    TODO: might need to take state
-    TODO: add descriptions of the arguments of the class, i.e., name, description, arguments
-    """
-
-    def __init__(self, name: str, description: str, arguments: list[ToolArgument]):
-        self.name = name
-        self.description = description
-        self.arguments = arguments
-
-    def validate_arguments(
-        self, provided_args: dict[str, Any]
-    ) -> tuple[bool, str | None]:
-        """Validate that all required arguments are provided with correct types"""
-        for arg in self.arguments:
-            if arg.required and arg.name not in provided_args:
-                return False, f"Missing required argument: {arg.name}"
-
-            if arg.name in provided_args:
-                value = provided_args[arg.name]
-
-                # Check choices if specified
-                if arg.choices is not None and value not in arg.choices:
-                    return (
-                        False,
-                        f"Invalid value for {arg.name}. Must be one of: {arg.choices}",
-                    )
-
-                try:
-                    # Basic type checking
-                    if arg.type == "int":
-                        int(value)
-                    elif arg.type == "float":
-                        float(value)
-                    elif arg.type == "bool":
-                        isinstance(value, bool)
-                except ValueError:
-                    return (
-                        False,
-                        f"Invalid type for argument {arg.name}. Expected {arg.type}",
-                    )
-
-        return True, None
-
-    def execute(self, **kwargs) -> str:
-        """Execute the tool functionality"""
-        raise NotImplementedError
-
-    def get_usage_guide(self) -> str:
-        """Generate a usage guide for the tool"""
-        args_desc = []
-        for arg in self.arguments:
-            required = (
-                "required" if arg.required else f"optional, default: {arg.default}"
-            )
-            args_desc.append(
-                f"- {arg.name} ({arg.type}, {required}): {arg.description}"
-            )
-
-        return f"""Tool: {self.name}
-Description: {self.description}
-Arguments:
-{chr(10).join(args_desc)}
-"""
-
-
-class ModalTool(Tool):
-    def __init__(
-        self,
-        modal_func: Callable,
-        name: str,
-        description: str,
-        arguments: list[ToolArgument],
-    ):
-        super().__init__(
-            name=name,
-            description=description,
-            arguments=arguments,
-        )
-        self._modal_func = modal_func
-
-    def execute(self, **kwargs):
-        return self._modal_func.remote(**kwargs)
 
 
 class Environment(ABC):
